subtask2/st2_requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.2
attrs==21.4.0
beautifulsoup4==4.11.1
bleach==5.0.1
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cycler==0.11.0
datasets==2.3.2
defusedxml==0.7.1
dill==0.3.5.1
fastjsonschema==2.16.1
filelock==3.7.1
fonttools==4.34.4
frozenlist==1.3.0
fsspec==2022.5.0
huggingface-hub==0.8.1
hyperopt==0.2.7
idna==3.3
ipython-genutils==0.2.0
ipywidgets==7.7.1
Jinja2==3.1.2
joblib==1.1.0
jsonlines==3.1.0
jsonschema==4.7.2
jupyter==1.0.0
jupyter-console==6.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.5.2
mistune==0.8.4
multidict==6.0.2
multiprocess==0.70.13
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nltk==3.7
notebook==6.4.12
numpy==1.23.1
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
Pillow==9.2.0
progressbar==2.5
prometheus-client==0.14.1
pyaml==21.10.1
pyarrow==8.0.0
pycparser==2.21
pymongo==3.12.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytz==2022.1
PyYAML==6.0
qtconsole==5.3.1
QtPy==2.1.0
regex==2022.7.9
requests==2.28.1
responses==0.18.0
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
seqeval==1.2.2
Send2Trash==1.8.0
sklearn==0.0
soupsieve==2.3.2.post1
terminado==0.15.0
threadpoolctl==3.1.0
tinycss2==1.1.1
tokenizers==0.12.1
torch==1.9.1
tqdm==4.64.0
traitlets==5.3.0
transformers==4.20.1
typing_extensions==4.3.0
urllib3==1.26.10
webencodings==0.5.1
widgetsnbextension==3.6.1
xxhash==3.0.0
yarl==1.7.2
